Move order coupon into order namespace
